+++ src/error/invalid-argument-error.ts
@@ -1,0 +1,12 @@
+import { ErrorExtensions, ExtendedError } from "../dto";
+
+import { appConfig } from "./../config";
+
+export class InvalidArgumentError extends ExtendedError<ErrorExtensions> {
+  constructor(message: string) {
+    super(message);
+    this.extensions = {
+      code: appConfig().errorCodePrefix + "002",
+    };
+  }
+}