+++ src/error/invalid-reset-password-request.error.ts
@@ -1,6 +1,13 @@
-export class InvalidResetPasswordRequestError extends Error {
+import { ErrorExtensions, ExtendedError } from "../dto";
+
+import { appConfig } from "../config";
+
+export class InvalidResetPasswordRequestError extends ExtendedError<ErrorExtensions> {
   constructor() {
     super();
     this.message = "Invalid reset password request.";
+    this.extensions = {
+      code: appConfig().errorCodePrefix + "005",
+    };
   }
 }