+++ src/error/undefined-service.error.ts
@@ -1,6 +1,13 @@
-export class UndefinedServiceError extends Error {
+import { ErrorExtensions, ExtendedError } from "../dto";
+
+import { appConfig } from "../config";
+
+export class UndefinedServiceError extends ExtendedError<ErrorExtensions> {
   constructor(service: string, serviceType: string) {
     super();
     this.message = `${serviceType} service is not defined for ${service}.`;
+    this.extensions = {
+      code: appConfig().errorCodePrefix + "007",
+    };
   }
 }