+++ src/login/login.module.ts
@@ -1,5 +1,6 @@
 import { LoginController } from "./login.controller";
 import { LoginOtpEmailController } from "./login-otp-email.controller";
+import { LoginOtpPhoneController } from "./login-otp-phone.controller";
 import { LoginProfile } from "./mapping-profiles/login.mapping-profile";
 import { Module } from "@nestjs/common";
 import { OtpModule } from "../otp/otp.module";
@@ -13,6 +14,7 @@ import { SendLoginOtpPhoneController } from "./send-login-otp-phone.controller";
   controllers: [
     LoginController,
     LoginOtpEmailController,
+    LoginOtpPhoneController,
     SendLoginOtpEmailController,
     SendLoginOtpPhoneController,
   ],