+++ src/login/send-login-otp-email.controller.ts
@@ -1,10 +1,4 @@
-import {
-  Inject,
-  Controller,
-  Post,
-  Body,
-  UnauthorizedException,
-} from "@nestjs/common";
+import { Inject, Controller, Post, Body } from "@nestjs/common";
 import { UserService } from "../user/user.service";
 import { ApiSecurity, ApiTags } from "@nestjs/swagger";
 import { SendLoginOtpEmailRequest } from "./dto/send-login-otp-email-request.dto";
@@ -53,7 +47,9 @@ export class SendLoginOtpEmailController {
       };
     } catch (error) {
       if (error instanceof InvalidCredentialsError)
-        throw new UnauthorizedException(null, { cause: error });
+        return this.otpService.createFakeOtpResult();
+
+      throw error;
     }
   }
 }