+++ test/fixtures/email/index.ts
@@ -1,2 +1,3 @@
 export * from "./email.fixture";
 export * from "./smtp-email-config.fixture";
+export * from "./sendgrid-email-config.fixture";