+++ CHANGELOG.md
@@ -2,6 +2,63 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.3.0](https://github.com/BrewInteractive/authentication-service-nestjs/compare/v2.2.0...v2.3.0) (2024-05-27)
+
+
+### Features
+
+* Generate random OTP result when sending OTP ([f4b8322](https://github.com/BrewInteractive/authentication-service-nestjs/commit/f4b832241c74f6ea605f1d9b211e814695472f13))
+* Remove unused import in otp.service.ts ([3db4e8e](https://github.com/BrewInteractive/authentication-service-nestjs/commit/3db4e8ef02b96866c2e728561827317ab222646a))
+
+## [2.2.0](https://github.com/BrewInteractive/authentication-service-nestjs/compare/v2.1.0...v2.2.0) (2024-05-27)
+
+
+### Features
+
+* Add fake OTP result for user not found ([f0a598f](https://github.com/BrewInteractive/authentication-service-nestjs/commit/f0a598f89de973ade2bc43103fe17f2f1642b2cc))
+* Add fake OTP result for user not found ([430ce04](https://github.com/BrewInteractive/authentication-service-nestjs/commit/430ce048c47debc833d9b8aafd65ea08d4f8f590))
+* Add faker library for generating fake OTP results ([3fcbab6](https://github.com/BrewInteractive/authentication-service-nestjs/commit/3fcbab6e08ab953ef168bc2742084760c54b4a72))
+* Add faker library for generating fake OTP results ([736fed4](https://github.com/BrewInteractive/authentication-service-nestjs/commit/736fed482e000aa8e40cfec84527fff4ac2f49df))
+* Fix error handling in SendLoginOtpEmailController ([4a8579d](https://github.com/BrewInteractive/authentication-service-nestjs/commit/4a8579d6dc461421625b802af6d4a23a2e1a3644))
+* Fix error handling in SendLoginOtpEmailController ([618be7a](https://github.com/BrewInteractive/authentication-service-nestjs/commit/618be7a82ef38e7dc59ff61908f635bafdff578e))
+* Handle invalid credentials in SendLoginOtpEmailController ([3b5f639](https://github.com/BrewInteractive/authentication-service-nestjs/commit/3b5f639c116126b27ecec726a2ee0b438b7136f0))
+* Handle invalid credentials in SendLoginOtpEmailController ([971eb4e](https://github.com/BrewInteractive/authentication-service-nestjs/commit/971eb4edfbb3b09b67f1a2fcad8494f3fdea4a37))
+* Refactor error classes to extend ExtendedError ([3163c54](https://github.com/BrewInteractive/authentication-service-nestjs/commit/3163c54e5c9cb15a32876c38a230d90ee0ed0aff))
+* Refactor OTP service to support phone OTPs and create `createPhoneOtpAsync` method ([7c704c2](https://github.com/BrewInteractive/authentication-service-nestjs/commit/7c704c27c4a308a48150b624ed605ba5c089de3b))
+* Refactor SendLoginOtpEmailController to handle user not found scenario ([b9d1e68](https://github.com/BrewInteractive/authentication-service-nestjs/commit/b9d1e68e658f1b6965481b1281464d18e380c8e3))
+* Refactor SendLoginOtpEmailController to handle user not found scenario ([fbe1e9c](https://github.com/BrewInteractive/authentication-service-nestjs/commit/fbe1e9c9d80b9a9ee4f6b6924b499d8fe5624454))
+* Remove duplicate import in otp.service.ts ([7cbef3e](https://github.com/BrewInteractive/authentication-service-nestjs/commit/7cbef3e7ecfe8f80b700effc8e6bd73f239b28f9))
+* Remove unused imports in otp.service.spec.ts ([0cb982c](https://github.com/BrewInteractive/authentication-service-nestjs/commit/0cb982cba5d78aa9aa7d045ec3904f7bc700b6f3))
+
+## [2.1.0](https://github.com/BrewInteractive/authentication-service-nestjs/compare/v2.0.1...v2.1.0) (2024-05-24)
+
+
+### Features
+
+* Add login OTP SMS template and service method ([543eedc](https://github.com/BrewInteractive/authentication-service-nestjs/commit/543eedc86645de4c6b01b921f0ae28c350f4cd9d))
+* Add MutluCell SMS service configuration ([d1a6a24](https://github.com/BrewInteractive/authentication-service-nestjs/commit/d1a6a245f7166cc25e2143d6b11c0954478e1c9b))
+* add MutluCellSmsService implementation ([3813021](https://github.com/BrewInteractive/authentication-service-nestjs/commit/3813021ac2d98eefdbf2aead922334f5254a0978))
+* add otp value generator ([4ff36c0](https://github.com/BrewInteractive/authentication-service-nestjs/commit/4ff36c0d4f0f27d8e6c191e17a2505877d46d3d6))
+* add phone number and country code to users table. ([357bebe](https://github.com/BrewInteractive/authentication-service-nestjs/commit/357bebea4177ef60f89d76199518611bd3e7948a))
+* Add phone number and country code validation to getUserAsync method ([7830eb1](https://github.com/BrewInteractive/authentication-service-nestjs/commit/7830eb1203132f6c091928d587af30ff9542e861))
+* Add phone OTP created event ([cc39ad4](https://github.com/BrewInteractive/authentication-service-nestjs/commit/cc39ad4e621ea8a8791f512b9dfc50f83a7702cb))
+* add sms service ([ac89f52](https://github.com/BrewInteractive/authentication-service-nestjs/commit/ac89f52819523d2cf45b51d2faf77dcda99f4bfe))
+* create custom validation pipe and its tests. ([b4d64e9](https://github.com/BrewInteractive/authentication-service-nestjs/commit/b4d64e94ec35e374dec2f3b98587b2a817882052))
+* expire otp function is added to otp service and used in login otp email controller. ([46a6d34](https://github.com/BrewInteractive/authentication-service-nestjs/commit/46a6d345a0f2f918ce1197c6f563c0d9dd5ab30e))
+* Improve error message in getUserAsync method ([4b582b7](https://github.com/BrewInteractive/authentication-service-nestjs/commit/4b582b75143fd742c7d3ed3f3c06ee657658bd90))
+* refactor create otp function to work with phone channel and create createPhoneOtpAsync method. ([057716d](https://github.com/BrewInteractive/authentication-service-nestjs/commit/057716d4737d953d264363e74836c3650970f39f))
+* Refactor getUserAsync method to include all user arguments ([9f010f4](https://github.com/BrewInteractive/authentication-service-nestjs/commit/9f010f4d9a35c4e6383eeafd628244847e4e100d))
+
+
+### Bug Fixes
+
+* edit validation error fields message. ([3f762c6](https://github.com/BrewInteractive/authentication-service-nestjs/commit/3f762c634ab4f74ac3daea4ad4391854fadd3a2b))
+* fix test case in controller login otp email. ([3aa0410](https://github.com/BrewInteractive/authentication-service-nestjs/commit/3aa0410b5237e02ebe5c21bceb6f9575ee6365d4))
+* fix typo in field name phoneNumber. ([0987b38](https://github.com/BrewInteractive/authentication-service-nestjs/commit/0987b385b7086797dce6c65825ff839a87b831c6))
+* fix validation error extention test case. ([c58d108](https://github.com/BrewInteractive/authentication-service-nestjs/commit/c58d108b2ab7773876b692b1c39d2eaf1be2692a))
+* remove html files and add folder to gitignore. ([a998bba](https://github.com/BrewInteractive/authentication-service-nestjs/commit/a998bba029d8a6bc3e09e5e635e4089e58b09c49))
+* remove unused import. ([134ea58](https://github.com/BrewInteractive/authentication-service-nestjs/commit/134ea583638fa5e940901affa032bc13be0f5d18))
+
 ### [2.0.1](https://github.com/BrewInteractive/authentication-service-nestjs/compare/v2.0.0...v2.0.1) (2024-05-21)